+++ hutool-core/src/main/java/cn/hutool/core/convert/Convert.java
@@ -1033,80 +1033,10 @@ public class Convert {
 	 *
 	 * @param chineseMoneyAmount 中文大写数字金额
 	 * @return 返回结果以元为单位的BigDecimal类型数字
+	 * @since 5.8.5
 	 */
-	public static BigDecimal chineseMoneyAmount2Number(String chineseMoneyAmount){
-		if(StrUtil.isBlank(chineseMoneyAmount)){
-			return null;
-		}
-
-		int yi = chineseMoneyAmount.indexOf("元");
-		if(yi == -1){
-			yi = chineseMoneyAmount.indexOf("圆");
-		}
-		int ji = chineseMoneyAmount.indexOf("角");
-		int fi = chineseMoneyAmount.indexOf("分");
-
-		/*
-		 * 先找到单位为元的数字
-		 */
-		String yStr = null;
-		if(yi > 0) {
-			yStr = chineseMoneyAmount.substring(0, yi);
-		}
-
-		/*
-		 * 再找到单位为角的数字
-		 */
-		String jStr = null;
-		if(ji > 0){
-			if(yi >= 0){
-				//前面有元,角肯定要在元后面
-				if(ji > yi){
-					jStr = chineseMoneyAmount.substring(yi+1, ji);
-				}
-			}else{
-				//没有元，只有角
-				jStr = chineseMoneyAmount.substring(0, ji);
-			}
-		}
-
-		/*
-		 * 再找到单位为分的数字
-		 */
-		String fStr = null;
-		if(fi > 0){
-			if(ji >= 0){
-				//有角，分肯定在角后面
-				if(fi > ji){
-					fStr = chineseMoneyAmount.substring(ji+1, fi);
-				}
-			}else if(yi > 0){
-				//没有角，有元，那就坐元后面找
-				if(fi > yi){
-					fStr = chineseMoneyAmount.substring(yi+1, fi);
-				}
-			}else {
-				//没有元、角，只有分
-				fStr = chineseMoneyAmount.substring(0, fi);
-			}
-		}
-
-		//元、角、分
-		int y = 0, j = 0, f = 0;
-		if(StrUtil.isNotBlank(yStr)) {
-			y = NumberChineseFormatter.chineseToNumber(yStr);
-		}
-		if(StrUtil.isNotBlank(jStr)){
-			j = NumberChineseFormatter.chineseToNumber(jStr);
-		}
-		if(StrUtil.isNotBlank(fStr)){
-			f = NumberChineseFormatter.chineseToNumber(fStr);
-		}
-
-		BigDecimal amount = new BigDecimal(y);
-		amount = amount.add(BigDecimal.valueOf(j).divide(BigDecimal.TEN));
-		amount = amount.add(BigDecimal.valueOf(f).divide(BigDecimal.valueOf(100)));
-		return amount;
+	public static BigDecimal chineseMoneyToNumber(String chineseMoneyAmount){
+		return NumberChineseFormatter.chineseMoneyToNumber(chineseMoneyAmount);
 	}
 
 	// -------------------------------------------------------------------------- 数字转换